+++ rayk-server/src/main/java/com/rayk/health/goldbean/vo/GoldBeanLedgerVo.java
@@ -1,12 +1,13 @@
 package com.rayk.health.goldbean.vo;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 public record GoldBeanLedgerVo(
         String id,
         String bucket,
         String direction,
-        long amount,
+        BigDecimal amount,
         String eventType,
         String description,
         LocalDateTime createdAt) {}
